+++ frontend/src/App.tsx
@@ -2,6 +2,7 @@ import { Routes, Route } from 'react-router-dom'
 import HomePage from './pages/HomePage'
 import SongsPage from './pages/SongsPage'
 import SongDetailPage from './pages/SongDetailPage'
+import SettingsPage from './pages/SettingsPage'
 
 function App() {
   return (
@@ -9,6 +10,7 @@ function App() {
       <Route path="/" element={<HomePage />} />
       <Route path="/songs" element={<SongsPage />} />
       <Route path="/song/:id" element={<SongDetailPage />} />
+      <Route path="/settings" element={<SettingsPage />} />
     </Routes>
   )
 }